+++ CriticalMass/Ride.swift
@@ -28,21 +28,3 @@ extension Ride {
         return "\(titleWithoutDate)\n\(dateTime.humanReadableDate()) - \(dateTime.humanReadableTime())"
     }
 }
-
-private extension String {
-    func removedDatePattern() -> String {
-        // assuming the date is formatted like dd.MM.yyyy. Also removes the whitespace before the date
-        let pattern = " \\d{1,2}.\\d{1,2}.\\d{4}$"
-        return removedRegexMatches(pattern: pattern)
-    }
-
-    private func removedRegexMatches(pattern: String, replaceWith template: String = "") -> String {
-        do {
-            let regex = try NSRegularExpression(pattern: pattern, options: NSRegularExpression.Options.caseInsensitive)
-            let range = NSRange(location: 0, length: count)
-            return regex.stringByReplacingMatches(in: self, options: [], range: range, withTemplate: template)
-        } catch {
-            return self
-        }
-    }
-}
